GaussianProcess

CalibrateEmulateSample.Emulators.GaussianProcessesPackage — Type

GaussianProcessesPackage

Type to dispatch which GP package to use

GPJL for GaussianProcesses.jl,

SKLJL for the ScikitLearn GaussianProcessRegressor

CalibrateEmulateSample.Emulators.PredictionType — Type

PredictionType

Predict type for GPJL in GaussianProcesses.jl

YType

FType latent function

CalibrateEmulateSample.Emulators.GaussianProcess — Type

GaussianProcess{FT<:AbstractFloat} <: MachineLearningTool

Structure holding training input and the fitted Gaussian Process Regression

models.

Fields

models

the Gaussian Process (GP) Regression model(s) that are fitted to the given input-data pairs

kernel

Kernel object

noise_learn

learn the noise with the White Noise kernel explicitly?

prediction_type

prediction type (y to predict the data, f to predict the latent function)

CalibrateEmulateSample.Emulators.build_models! — Function

function build_models!(
    gp::GaussianProcess{package},
    input_output_pairs)

method to build gaussian process models based on the package

CalibrateEmulateSample.Emulators.optimize_hyperparameters! — Method

function optimize_hyperparameters!(gp::GaussianProcess{package})

optimize Gaussian Process hyperparameters using in-build package method