GaussianProcessEmulator

CalibrateEmulateSample.GaussianProcessEmulator.GaussianProcessesPackage — Type

GaussianProcessesPackage

Type to dispatch which GP package to use

GPJL for GaussianProcesses.jl,

SKLJL for the ScikitLearn GaussianProcessRegressor

CalibrateEmulateSample.GaussianProcessEmulator.PredictionType — Type

PredictionType

Predict type for GPJL in GaussianProcesses.jl

YType

FType latent function

CalibrateEmulateSample.GaussianProcessEmulator.GaussianProcess — Type

GaussianProcess{FT<:AbstractFloat}

Structure holding training input and the fitted Gaussian Process Regression

models.

Fields

input_output_pairs

training inputs and outputs, data stored in columns

input_mean

mean of input; 1 × input_dim

sqrt_inv_input_cov

square root of the inverse of the input covariance matrix; inputdim × inputdim

models

the Gaussian Process (GP) Regression model(s) that are fitted to the given input-data pairs

decomposition

the singular value decomposition of obsnoisecov, such that obsnoisecov = decomposition.U * Diagonal(decomposition.S) * decomposition.Vt.

normalized

whether to fit GP models on normalized inputs ((inputs - inputmean) * sqrtinvinputcov)

prediction_type

prediction type (y to predict the data, f to predict the latent function)

norm_factors

Standardization factors (characteristic values of the problem)

CalibrateEmulateSample.GaussianProcessEmulator.predict — Function

predict(gp::GaussianProcess{FT, GPJL}, new_inputs::Array{FT, 2}, transform_to_real::Bool=false) where {FT} = predict(gp, new_inputs, gp.prediction_type)

Evaluate the GP model(s) at new inputs.

gp - a GaussianProcess

new_inputs - inputs for which GP model(s) is/are evaluated; inputdim × Nsamples

Returns the predicted mean(s) and covariance(s) at the input points. Means: matrix of size outputdim × Nsamples Covariances: vector of length Nsamples, each element is a matrix of size outputdim × outputdim. If the output is 1-dimensional, a 1 × Nsamples array of scalar variances is returned rather than a vector of 1x1 matrices.

Note: If gp.normalized == true, the new inputs are normalized prior to the prediction

CalibrateEmulateSample.GaussianProcessEmulator.svd_transform — Function

svdtransform(data::Array{FT, 2}, obsnoise_cov::Union{Array{FT, 2}, Nothing}) where {FT}

Apply a singular value decomposition (SVD) to the data

data - GP training data/targets; outputdim × Nsamples

obs_noise_cov - covariance of observational noise

Returns the transformed data and the decomposition, which is a matrix factorization of type LinearAlgebra.SVD.

Note: If F::SVD is the factorization object, U, S, V and Vt can be obtained via F.U, F.S, F.V and F.Vt, such that A = U * Diagonal(S) * Vt. The singular values in S are sorted in descending order.

CalibrateEmulateSample.GaussianProcessEmulator.svd_reverse_transform_mean_cov — Function

svdreversetransformmeancov(μ::Array{FT, 2}, σ2::{Array{FT, 2}, decomposition::SVD) where {FT}

Transform the mean and covariance back to the original (correlated) coordinate system

μ - predicted mean; outputdim × Npredicted_points

σ2 - predicted variance; outputdim × Npredicted_points

Returns the transformed mean (outputdim × Npredictedpoints) and variance. Note that transforming the variance back to the original coordinate system results in non-zero off-diagonal elements, so instead of just returning the elements on the main diagonal (i.e., the variances), we return the full covariance at each point, as a vector of length Npredictedpoints, where each element is a matrix of size outputdim × output_dim