MarkovChainMonteCarlo

CalibrateEmulateSample.MarkovChainMonteCarlo.MCMC — Type

mutable struct MCMC{FT<:AbstractFloat, IT<:Int64}

Structure to organize MCMC parameters and data.

Fields

obs_sample::AbstractVector{FT} where FT<:AbstractFloat

A single sample from the observations. Can e.g. be picked from an Obs struct using get_obs_sample.

obs_noise_cov::Union{AbstractMatrix{FT}, LinearAlgebra.UniformScaling{FT}} where FT<:AbstractFloat

Covariance of the observational noise.

prior::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution

Array of length N_parameters with the parameters' prior distributions.

step::AbstractFloat

MCMC step size.

burnin::Int64

Number of MCMC steps that are considered burnin.

param::AbstractVector{FT} where FT<:AbstractFloat

The current parameters.

posterior::AbstractMatrix{FT} where FT<:AbstractFloat

Array of accepted MCMC parameter samples (param_dim × n_samples). The histogram of these samples gives an approximation of the posterior distribution of the parameters.

log_posterior::Union{Nothing, FT} where FT<:AbstractFloat

The current value of the logarithm of the posterior (= log_likelihood + log_prior of the current parameters).

iter::Int64

Iteration/step of the MCMC.

accept::Int64

Number of accepted proposals.

algtype::String

MCMC algorithm to use. Currently implemented: 'rmw' (random walk Metropolis), 'pCN' (preconditioned Crank-Nicholson).

rng::Random.AbstractRNG

Random number generator object (algorithm + seed) used for sampling and noise, for reproducibility.

CalibrateEmulateSample.MarkovChainMonteCarlo.MCMC — Method

MCMC(obs_sample::AbstractArray{FT<:AbstractFloat, 1}, obs_noise_cov::Union{AbstractArray{FT<:AbstractFloat, 2}, LinearAlgebra.UniformScaling{FT<:AbstractFloat}}, prior::EnsembleKalmanProcesses.ParameterDistributions.ParameterDistribution, step::AbstractFloat, param_init::AbstractArray{FT<:AbstractFloat, 1}, max_iter::Int64, algtype::String, burnin::Int64; svdflag, standardize, norm_factor, truncate_svd, rng) -> CalibrateEmulateSample.MarkovChainMonteCarlo.MCMC{_A, Int64} where _A<:AbstractFloat

Constructor for MCMC.

max_iter - The number of MCMC steps to perform (e.g., 100_000).