Emulators

CalibrateEmulateSample.Emulators.Decomposition — Type

Decomposition{<:AbstractFloat, <:Int}

Struct of SVD decomposition, containing (V,S,Vt), and the size of S, N.

CalibrateEmulateSample.Emulators.Emulator — Type

Emulator

Structure used to represent a general emulator:

CalibrateEmulateSample.Emulators.optimize_hyperparameters! — Method

function optimize_hyperparameters!(emulator::Emulator)

optimize the hyperparameters in the machine learning tool

CalibrateEmulateSample.Emulators.predict — Function

function predict(gp::GaussianProcess{package}, new_inputs::Array{FT, 2})

predict means and covariances in decorrelated output space using gaussian process models

function predict(emulator::Emulator, new_inputs; transform_to_real=false)

makes a prediction using the emulator on new inputs (each new inputs given as data columns), default is to predict in the decorrelated space

CalibrateEmulateSample.Emulators.normalize — Function

function normalize(emulator::Emulator, inputs)

normalize the input data, with a normalizing function

function normalize(inputs, input_mean, sqrt_inv_input_cov)

normalize with the empirical Gaussian distribution of points

CalibrateEmulateSample.Emulators.standardize — Function

function standardize(outputs, output_cov, factors)

standardize with a vector of factors (size equal to output dimension)

CalibrateEmulateSample.Emulators.reverse_standardize — Function

function reverse_standardize(emulator::Emulator, outputs, output_cov)

reverse a previous standardization with the stored vector of factors (size equal to output dimension)

CalibrateEmulateSample.Emulators.svd_transform — Function

svdtransform(data::Array{FT, 2}, obsnoise_cov::Union{Array{FT, 2}, Nothing}) where {FT}

Apply a singular value decomposition (SVD) to the data

data - GP training data/targets; outputdim × Nsamples

obs_noise_cov - covariance of observational noise

Returns the transformed data and the decomposition, which is a matrix factorization of type LinearAlgebra.SVD.

Note: If F::SVD is the factorization object, U, S, V and Vt can be obtained via F.U, F.S, F.V and F.Vt, such that A = U * Diagonal(S) * Vt. The singular values in S are sorted in descending order.

function svdtransform( data::Vector{FT}, obsnoisecov::Union{Array{FT, 2}, Nothing}; truncatesvd::FT=1.0) where {FT}

CalibrateEmulateSample.Emulators.svd_reverse_transform_mean_cov — Function

svdreversetransformmeancov(μ::Array{FT, 2}, σ2::{Array{FT, 2}, decomposition::SVD) where {FT}

Transform the mean and covariance back to the original (correlated) coordinate system

μ - predicted mean; outputdim × Npredicted_points

σ2 - predicted variance; outputdim × Npredicted_points

Returns the transformed mean (outputdim × Npredictedpoints) and variance. Note that transforming the variance back to the original coordinate system results in non-zero off-diagonal elements, so instead of just returning the elements on the main diagonal (i.e., the variances), we return the full covariance at each point, as a vector of length Npredictedpoints, where each element is a matrix of size outputdim × output_dim