GaussianProcess

CalibrateEmulateSample.Emulators.GaussianProcessesPackage — Type

abstract type GaussianProcessesPackage

Type to dispatch which GP package to use:

GPJL for GaussianProcesses.jl,

SKLJL for the ScikitLearn GaussianProcessRegressor.

CalibrateEmulateSample.Emulators.PredictionType — Type

abstract type PredictionType

Predict type for GPJL in GaussianProcesses.jl:

YType

FType latent function.

CalibrateEmulateSample.Emulators.GaussianProcess — Type

struct GaussianProcess{GPPackage, FT} <: CalibrateEmulateSample.Emulators.MachineLearningTool

Structure holding training input and the fitted Gaussian process regression models.

Fields

models::Vector{Union{Nothing, PyCall.PyObject, GaussianProcesses.GPE}}: The Gaussian Process (GP) Regression model(s) that are fitted to the given input-data pairs.

kernel::Union{Nothing, var"#s12", var"#s13"} where {var"#s12"<:GaussianProcesses.Kernel, var"#s13"<:PyCall.PyObject}: Kernel object.

noise_learn::Bool: Learn the noise with the White Noise kernel explicitly?

alg_reg_noise::Any: Additional observational or regularization noise in used in GP algorithms

prediction_type::CalibrateEmulateSample.Emulators.PredictionType: Prediction type (y to predict the data, f to predict the latent function).

CalibrateEmulateSample.Emulators.build_models! — Method

build_models!(
    gp::CalibrateEmulateSample.Emulators.GaussianProcess{CalibrateEmulateSample.Emulators.GPJL},
    input_output_pairs::EnsembleKalmanProcesses.DataContainers.PairedDataContainer{FT<:AbstractFloat}
)

Method to build Gaussian process models based on the package.

CalibrateEmulateSample.Emulators.optimize_hyperparameters! — Method

optimize_hyperparameters!(
    gp::CalibrateEmulateSample.Emulators.GaussianProcess{CalibrateEmulateSample.Emulators.GPJL},
    args...;
    kwargs...
)

Optimize Gaussian process hyperparameters using in-build package method.

Warning: if one uses GPJL() and wishes to modify positional arguments. The first positional argument must be the Optim method (default LBGFS()).

GaussianProcesses.predict — Method

predict(
    gp::CalibrateEmulateSample.Emulators.GaussianProcess{CalibrateEmulateSample.Emulators.GPJL},
    new_inputs::AbstractArray{FT<:AbstractFloat, 2}
) -> Tuple{Any, Any}

Predict means and covariances in decorrelated output space using Gaussian process models.