Utilities

CalibrateEmulateSample.Utilities.AffineScaler — Type

struct AffineScaler{T, VV<:(AbstractVector)} <: CalibrateEmulateSample.Utilities.DataContainerProcessor

The AffineScaler{T} will create an encoding of the data_container via affine transformations.

Different methods T will build different transformations:

quartile_scale : creates QuartileScaling,

minmax_scale : creates MinMaxScaling

zscore_scale : creates ZScoreScaling

and are accessed with get_type

CalibrateEmulateSample.Utilities.CanonicalCorrelation — Type

struct CanonicalCorrelation{VV1, VV2, VV3, FT, VV4} <: CalibrateEmulateSample.Utilities.PairedDataContainerProcessor

Uses both input and output data to learn a subspace of maximal correlation between inputs and outputs. The subspace for a pair (X,Y) will be of size minimum(rank(X),rank(Y)), computed using SVD-based method e.g. See e.g., https://numerical.recipes/whp/notes/CanonCorrBySVD.pdf

Preferred construction is with the canonical_correlation method

Fields

data_mean::Any: storage for the input or output data mean

encoder_mat::Any: the encoding matrix of input or output canonical correlations

decoder_mat::Any: the decoding matrix of input or output canonical correlations

retain_var::Any: the fraction of variance to be retained after truncating singular values (1 implies no truncation)

apply_to::Any: Stores whether this is an input or output encoder (vector with string "in" or "out")

CalibrateEmulateSample.Utilities.Decorrelater — Type

struct Decorrelater{VV1, VV2, VV3, FT} <: CalibrateEmulateSample.Utilities.DataContainerProcessor

Decorrelate the data via an SVD decomposition using a structurematrix (`priorcovfor inputs,obsnoisecovfor outputs), with optional truncation of singular vectors corresponding to largest singular values. The structure matrix will also become exactlyI` after processing.

Similar to the Standardizer, except that the Standardizer uses the estimated covariance of the data to decorrelate in-place of the structure matrix. There, the data samples will have sample mean 0 and covariance I after processing.

In the Decorrelater the user can do a combined approach where one uses cov(data) + structure matrix for decorrelation with the add_estimated_data_cov=true

Preferred construction is with the decorrelate method

Fields

data_mean::Any: storage for the data mean

encoder_mat::Any: the (possibly-encoded) structure matrix - provided by the user

decoder_mat::Any: the inverse of the (possibly-encoded) structure matrix

retain_var::Any: the fraction of variance to be retained after truncating singular values (1 implies no truncation)

add_estimated_cov::Bool: If true, add the estimated cov(data) to the structure matrix for to modify the decorrelation

CalibrateEmulateSample.Utilities.Standardizer — Type

struct Standardizer{VV1<:(AbstractVector), VV2<:(AbstractVector), VV3<:(AbstractVector)} <: CalibrateEmulateSample.Utilities.DataContainerProcessor

Standardizes the data to a multivariate N(0,I) distribution via (xcov)^{-1/2}*(x-x_mean), along with rank reduction if data is low rank, or if the user provides a rank. This guarantees that the data samples will have sample mean 0 and covariance I after processing.

Similar to the Decorrelater, but there the structure matrices will directly be used to perform decorrelation, and will become I after processing. In that case, the data samples may not have covariance I after processing.

Preferred construction is using the standardize method.

Fields

rank::Int64: user-provided rank for additional truncation of the space. used if rank < rank(data_cov)

data_mean::AbstractVector: storage for the data mean

encoder_mat::AbstractVector: storage for wide-matrix representing data_cov^{-1/2}

decoder_mat::AbstractVector: storage for tall-matrix representing data_cov^{1/2}

CalibrateEmulateSample.Utilities.canonical_correlation — Method

canonical_correlation(
;
    retain_var
) -> CalibrateEmulateSample.Utilities.CanonicalCorrelation{Vector{Any}, Vector{Any}, Vector{Any}, Float64, Vector{AbstractString}}

Constructs the CanonicalCorrelation struct. Can optionally provide the keyword

retain_var[=1.0]: to project onto the leading singular vectors (of the input-output product) such that retain_var variance is retained.

CalibrateEmulateSample.Utilities.create_encoder_schedule — Method

Create size-1 encoder schedule with a tuple of (DataProcessor1(...), apply_to) with apply_to = "in", "out" or "in_and_out".

CalibrateEmulateSample.Utilities.create_encoder_schedule — Method

create_encoder_schedule(
    schedule_in::AbstractVector
) -> Vector{Any}

Create a flatter encoder schedule for the from the user's proposed schedule of the form:

enc_schedule = [
    (DataProcessor1(...), "in"), 
    (DataProcessor2(...), "out"), 
    (PairedDataProcessor3(...), "in"), 
    (DataProcessor4(...), "in_and_out"), 
]

This function creates the encoder scheduler that is also machine readable

enc_schedule = [
    (DataProcessor1(...), x -> get_inputs(x), "in"), 
    (DataProcessor2(...), x -> get_outputs(x), "out"), 
    (DataProcessor2(...), x -> get_outputs(x), "out"),
    (PairedDataProcessor3(...), x -> (get_outputs(x), get_outputs(x)), "in"), 
    (DataProcessor4(...), x -> get_inputs(x), "in"),
    (DataProcessor4(...), x -> get_outputs(x), "out"), 
]

and the decoder schedule is a copy of the encoder schedule reversed (and processors copied)

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    dcp::CalibrateEmulateSample.Utilities.DataContainerProcessor,
    data::AbstractMatrix,
    apply_to::AbstractString
) -> Any

decodes the columns-are-data matrix with the processor.

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    dcp::CalibrateEmulateSample.Utilities.PairedDataContainerProcessor,
    data,
    apply_to::AbstractString
) -> Any

decodes the input or output dat (pair of columns-are-data matrices) a with the processor, based on apply_to.

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    as::CalibrateEmulateSample.Utilities.AffineScaler,
    data::AbstractMatrix
) -> AbstractMatrix

Apply the AffineScaler decoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation,
    data::AbstractMatrix
) -> Any

Apply the CanonicalCorrelation decoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    dd::CalibrateEmulateSample.Utilities.Decorrelater,
    data::AbstractMatrix
) -> Any

Apply the Decorrelater decoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.decode_data — Method

decode_data(
    ss::CalibrateEmulateSample.Utilities.Standardizer,
    data::AbstractMatrix
) -> Any

Apply the Standardizer decoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.decode_structure_matrix — Method

decode_structure_matrix(
    as::CalibrateEmulateSample.Utilities.AffineScaler,
    enc_structure_matrix::AbstractMatrix
) -> Any

Apply the AffineScaler decoder to a provided structure matrix

CalibrateEmulateSample.Utilities.decode_structure_matrix — Method

decode_structure_matrix(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation,
    enc_structure_matrix::AbstractMatrix
) -> Any

Apply the CanonicalCorrelation decoder to a provided structure matrix

CalibrateEmulateSample.Utilities.decode_structure_matrix — Method

decode_structure_matrix(
    dd::CalibrateEmulateSample.Utilities.Decorrelater,
    enc_structure_matrix::AbstractMatrix
) -> Any

Apply the Decorrelater decoder to a provided structure matrix

CalibrateEmulateSample.Utilities.decode_structure_matrix — Method

decode_structure_matrix(
    ss::CalibrateEmulateSample.Utilities.Standardizer,
    enc_structure_matrix::AbstractMatrix
) -> Any

Apply the Standardizer decoder to a provided structure matrix

CalibrateEmulateSample.Utilities.decode_with_schedule — Method

decode_with_schedule(
    encoder_schedule::AbstractVector,
    data_container::EnsembleKalmanProcesses.DataContainers.DataContainer,
    in_or_out::AbstractString
) -> EnsembleKalmanProcesses.DataContainers.DataContainer

Takes in an already initialized encoder schedule, and decodes a DataContainer, the in_or_out string indicates if the data is input "in" or output "out" data (and thus decoded differently)

CalibrateEmulateSample.Utilities.decode_with_schedule — Method

decode_with_schedule(
    encoder_schedule::AbstractVector,
    structure_matrix::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    in_or_out::AbstractString
) -> Any

Takes in an already initialized encoder schedule, and decodes a structure matrix, the in_or_out string indicates if the structure matrix is for input "in" or output "out" space (and thus decoded differently)

CalibrateEmulateSample.Utilities.decode_with_schedule — Method

decode_with_schedule(
    encoder_schedule::AbstractVector,
    io_pairs::EnsembleKalmanProcesses.DataContainers.PairedDataContainer,
    prior_cov_in::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    obs_noise_cov_in::Union{LinearAlgebra.UniformScaling, AbstractMatrix}
) -> Tuple{EnsembleKalmanProcesses.DataContainers.PairedDataContainer, Any, Any}

Takes in an already initialized encoder schedule, and decodes a DataContainer, and structure matrices with it, the in_or_out string indicates if the data is input "in" or output "out" data (and thus decoded differently)

CalibrateEmulateSample.Utilities.decorrelate — Method

decorrelate(
;
    retain_var,
    add_estimated_cov
) -> CalibrateEmulateSample.Utilities.Decorrelater{Vector{Any}, Vector{Any}, Vector{Any}, Float64}

Constructs the Decorrelater struct. Users can add optional keyword arguments:

retain_var[=1.0]: to project onto the leading singular vectors such that retain_var variance is retained

add_estimated_cov[=false]: to add the estimated covariance to the structure matrix, and use this summation to create the new subspace. (false uses just the structure matrix)

CalibrateEmulateSample.Utilities.encode_data — Method

encode_data(
    as::CalibrateEmulateSample.Utilities.AffineScaler,
    data::AbstractMatrix
) -> AbstractMatrix

Apply the AffineScaler encoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.encode_data — Method

encode_data(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation,
    data::AbstractMatrix
) -> Any

Apply the CanonicalCorrelation encoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.encode_data — Method

encode_data(
    dd::CalibrateEmulateSample.Utilities.Decorrelater,
    data::AbstractMatrix
) -> Any

Apply the Decorrelater encoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.encode_data — Method

encode_data(
    ss::CalibrateEmulateSample.Utilities.Standardizer,
    data::AbstractMatrix
) -> Any

Apply the Standardizer encoder, on a columns-are-data matrix

CalibrateEmulateSample.Utilities.encode_structure_matrix — Method

encode_structure_matrix(
    as::CalibrateEmulateSample.Utilities.AffineScaler,
    structure_matrix::AbstractMatrix
) -> Any

Apply the AffineScaler encoder to a provided structure matrix

CalibrateEmulateSample.Utilities.encode_structure_matrix — Method

encode_structure_matrix(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation,
    structure_matrix::AbstractMatrix
) -> Any

Apply the CanonicalCorrelation encoder to a provided structure matrix

CalibrateEmulateSample.Utilities.encode_structure_matrix — Method

encode_structure_matrix(
    dd::CalibrateEmulateSample.Utilities.Decorrelater,
    structure_matrix::AbstractMatrix
) -> Any

Apply the Decorrelater encoder to a provided structure matrix

CalibrateEmulateSample.Utilities.encode_structure_matrix — Method

encode_structure_matrix(
    ss::CalibrateEmulateSample.Utilities.Standardizer,
    structure_matrix::AbstractMatrix
) -> Any

Apply the Standardizer encoder to a provided structure matrix

CalibrateEmulateSample.Utilities.encode_with_schedule — Method

encode_with_schedule(
    encoder_schedule::AbstractVector,
    data_container::EnsembleKalmanProcesses.DataContainers.DataContainer,
    in_or_out::AbstractString
) -> EnsembleKalmanProcesses.DataContainers.DataContainer

Takes in an already initialized encoder schedule, and encodes a DataContainer, the in_or_out string indicates if the data is input "in" or output "out" data (and thus encoded differently)

CalibrateEmulateSample.Utilities.encode_with_schedule — Method

encode_with_schedule(
    encoder_schedule::AbstractVector,
    structure_matrix::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    in_or_out::AbstractString
) -> Any

Takes in an already initialized encoder schedule, and encodes a structure matrix, the in_or_out string indicates if the structure matrix is for input "in" or output "out" space (and thus encoded differently)

CalibrateEmulateSample.Utilities.encode_with_schedule — Method

encode_with_schedule(
    encoder_schedule::AbstractVector,
    io_pairs::EnsembleKalmanProcesses.DataContainers.PairedDataContainer,
    prior_cov_in::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    obs_noise_cov_in::Union{LinearAlgebra.UniformScaling, AbstractMatrix}
) -> Tuple{EnsembleKalmanProcesses.DataContainers.PairedDataContainer, Any, Any}

Takes in the created encoder schedule (See create_encoder_schedule), and initializes it, and encodes the paired data container, and structure matrices with it.

CalibrateEmulateSample.Utilities.get_add_estimated_cov — Method

get_add_estimated_cov(
    dd::CalibrateEmulateSample.Utilities.Decorrelater
) -> Bool

returns the add_estimated_cov field of the Decorrelater.

CalibrateEmulateSample.Utilities.get_apply_to — Method

get_apply_to(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation
) -> Any

returns the apply_to field of the CanonicalCorrelation.

CalibrateEmulateSample.Utilities.get_data_mean — Method

get_data_mean(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation
) -> Any

returns the data_mean field of the CanonicalCorrelation.

CalibrateEmulateSample.Utilities.get_data_mean — Method

get_data_mean(
    dd::CalibrateEmulateSample.Utilities.Decorrelater
) -> Any

returns the data_mean field of the Decorrelater.

CalibrateEmulateSample.Utilities.get_data_mean — Method

get_data_mean(
    ss::CalibrateEmulateSample.Utilities.Standardizer
) -> AbstractVector

returns the data_mean field of the Standardizer.

CalibrateEmulateSample.Utilities.get_decoder_mat — Method

get_decoder_mat(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation
) -> Any

returns the decoder_mat field of the CanonicalCorrelation.

CalibrateEmulateSample.Utilities.get_decoder_mat — Method

get_decoder_mat(
    dd::CalibrateEmulateSample.Utilities.Decorrelater
) -> Any

returns the decoder_mat field of the Decorrelater.

CalibrateEmulateSample.Utilities.get_decoder_mat — Method

get_decoder_mat(
    ss::CalibrateEmulateSample.Utilities.Standardizer
) -> AbstractVector

returns the decoder_mat field of the Standardizer.

CalibrateEmulateSample.Utilities.get_encoder_mat — Method

get_encoder_mat(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation
) -> Any

returns the encoder_mat field of the CanonicalCorrelation.

CalibrateEmulateSample.Utilities.get_encoder_mat — Method

get_encoder_mat(
    dd::CalibrateEmulateSample.Utilities.Decorrelater
) -> Any

returns the encoder_mat field of the Decorrelater.

CalibrateEmulateSample.Utilities.get_encoder_mat — Method

get_encoder_mat(
    ss::CalibrateEmulateSample.Utilities.Standardizer
) -> AbstractVector

returns the encoder_mat field of the Standardizer.

CalibrateEmulateSample.Utilities.get_rank — Method

get_rank(
    ss::CalibrateEmulateSample.Utilities.Standardizer
) -> Int64

returns the rank field of the Standardizer.

CalibrateEmulateSample.Utilities.get_retain_var — Method

get_retain_var(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation
) -> Any

returns the retain_var field of the CanonicalCorrelation.

CalibrateEmulateSample.Utilities.get_retain_var — Method

get_retain_var(
    dd::CalibrateEmulateSample.Utilities.Decorrelater
) -> Any

returns the retain_var field of the Decorrelater.

CalibrateEmulateSample.Utilities.get_scale — Method

get_scale(
    as::CalibrateEmulateSample.Utilities.AffineScaler
) -> AbstractVector

Gets the scale field of the AffineScaler

CalibrateEmulateSample.Utilities.get_shift — Method

get_shift(
    as::CalibrateEmulateSample.Utilities.AffineScaler
) -> AbstractVector

Gets the shift field of the AffineScaler

CalibrateEmulateSample.Utilities.get_training_points — Method

get_training_points(
    ekp::EnsembleKalmanProcesses.EnsembleKalmanProcess{FT, IT, P},
    train_iterations::Union{AbstractVector{IT}, IT} where IT
) -> EnsembleKalmanProcesses.DataContainers.PairedDataContainer

Extract the training points needed to train the Gaussian process regression.

ekp - EnsembleKalmanProcess holding the parameters and the data that were produced during the Ensemble Kalman (EK) process.

train_iterations - Number (or indices) EK layers/iterations to train on.

CalibrateEmulateSample.Utilities.get_type — Method

get_type(
    as::CalibrateEmulateSample.Utilities.AffineScaler{T, VV}
) -> Any

Gets the UnivariateAffineScaling type T

CalibrateEmulateSample.Utilities.initialize_and_encode_data! — Method

initialize_and_encode_data!(
    dcp::CalibrateEmulateSample.Utilities.DataContainerProcessor,
    data::AbstractMatrix,
    structure_mat::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    apply_to::AbstractString
) -> Any

Initializes the DataContainerProcessor encoder (often requires data, and structure matrices), then encodes the provided columns-are-data matrix

CalibrateEmulateSample.Utilities.initialize_and_encode_data! — Method

initialize_and_encode_data!(
    dcp::CalibrateEmulateSample.Utilities.PairedDataContainerProcessor,
    data,
    structure_mat::Union{LinearAlgebra.UniformScaling, AbstractMatrix},
    apply_to::AbstractString
) -> Any

Initializes the PairedDataContainerProcesser encoder (often requires input & output data, and structure matrices), then encodes either the input or output data (pair of columns-are-data matrices) based on apply_to.

CalibrateEmulateSample.Utilities.initialize_processor! — Method

initialize_processor!(
    cc::CalibrateEmulateSample.Utilities.CanonicalCorrelation,
    in_data::AbstractMatrix,
    out_data::AbstractMatrix,
    structure_matrix,
    apply_to::AbstractString
) -> Any

Computes and populates the data_mean, encoder_mat, decoder_mat and apply_to fields for the CanonicalCorrelation

CalibrateEmulateSample.Utilities.initialize_processor! — Method

initialize_processor!(
    as::CalibrateEmulateSample.Utilities.AffineScaler,
    data::AbstractMatrix,
    structure_matrix
) -> Any

Computes and populates the shift and scale fields for the AffineScaler

CalibrateEmulateSample.Utilities.initialize_processor! — Method

initialize_processor!(
    ss::CalibrateEmulateSample.Utilities.Standardizer,
    data::AbstractMatrix,
    structure_matrix
) -> Any

Computes and populates the data_mean and encoder_mat and decoder_mat fields for the Standardizer

CalibrateEmulateSample.Utilities.initialize_processor! — Method

initialize_processor!(
    dd::CalibrateEmulateSample.Utilities.Decorrelater,
    data::AbstractMatrix,
    structure_matrix::Union{LinearAlgebra.UniformScaling, AbstractMatrix}
) -> Any

Computes and populates the data_mean and encoder_mat and decoder_mat fields for the Decorrelater

CalibrateEmulateSample.Utilities.minmax_scale — Method

minmax_scale(

) -> CalibrateEmulateSample.Utilities.AffineScaler{CalibrateEmulateSample.Utilities.MinMaxScaling, Vector{Float64}}

Constructs AffineScaler{MinMaxScaling} processor. As part of an encoder schedule, this will apply the transform $\frac{x - \min(x)}{\max(x) - \min(x)}$ to each data dimension.

CalibrateEmulateSample.Utilities.quartile_scale — Method

quartile_scale(

) -> CalibrateEmulateSample.Utilities.AffineScaler{CalibrateEmulateSample.Utilities.QuartileScaling, Vector{Float64}}

Constructs AffineScaler{QuartileScaling} processor. As part of an encoder schedule, it will apply the transform $\frac{x - Q2(x)}{Q3(x) - Q1(x)}$ to each data dimension. Also known as "robust scaling"

CalibrateEmulateSample.Utilities.standardize — Method

standardize(
    rk::Int64
) -> CalibrateEmulateSample.Utilities.Standardizer{Vector{Float64}, Vector{Any}, Vector{Any}}

Constructs the Standardizer struct with a provided truncation to rank rk.

CalibrateEmulateSample.Utilities.standardize — Method

standardize(

) -> CalibrateEmulateSample.Utilities.Standardizer{Vector{Float64}, Vector{Any}, Vector{Any}}

Constructs the Standardizer struct with no truncation of rank

CalibrateEmulateSample.Utilities.zscore_scale — Method

zscore_scale(

) -> CalibrateEmulateSample.Utilities.AffineScaler{CalibrateEmulateSample.Utilities.ZScoreScaling, Vector{Float64}}

Constructs AffineScaler{ZScoreScaling} processor. As part of an encoder schedule, this will apply the transform $\frac{x-\mu}{\sigma}$, (where $x\sim N(\mu,\sigma)$), to each data dimension. For multivariate standardization, see Standardizer