Utilities

CalibrateEmulateSample.Utilities.get_obs_sample — Method

get_obs_sample(
    rng::Random.AbstractRNG,
    obs::EnsembleKalmanProcesses.Observations.Observation;
    rng_seed
) -> Any

Return a random sample from the observations, for use in the MCMC.

rng - optional RNG object used to pick random sample; defaults to Random.GLOBAL_RNG.

obs - Observation struct with the observations (extract will pick one of the sample observations to train).

rng_seed - optional kwarg; if provided, used to re-seed rng before sampling.

CalibrateEmulateSample.Utilities.get_training_points — Method

get_training_points(
    ekp::EnsembleKalmanProcesses.EnsembleKalmanProcess{FT, IT, P},
    train_iterations::Union{AbstractVector{IT}, IT} where IT
) -> EnsembleKalmanProcesses.DataContainers.PairedDataContainer

Extract the training points needed to train the Gaussian process regression.

ekp - EnsembleKalmanProcess holding the parameters and the data that were produced during the Ensemble Kalman (EK) process.

train_iterations - Number (or indices) EK layers/iterations to train on.